+++ src/app/app-routing.module.ts
@@ -15,7 +15,7 @@ const routes: Routes = [
   },
   {
     path: 'films',
-    component: FilmsComponent
+    component: FilmsComponent,
   },
   {
     path: 'people',
