+++ src/app/components/film/film.component.html
@@ -68,4 +68,5 @@
       >
     </div>
   </div>
+  <app-pagination></app-pagination>
 </div>
