+++ src/app/components/film/film.component.scss
@@ -1,0 +1,4 @@
+.flechas {
+    padding: 0;
+    margin: 0;
+}
