+++ src/app/components/pagination/pagination.component.html
@@ -1,0 +1,16 @@
+<div class="flechas mt-2">
+  <nav aria-label="Page navigation example">
+    <ul class="pagination d-flex justify-content-between fs-3">
+      <li class="page-item">
+        <button class="btn btn-outline-primary">
+          <i class="bi bi-arrow-left"></i>
+        </button>
+      </li>
+      <li class="page-item">
+        <button class="btn btn-outline-primary">
+          <i class="bi bi-arrow-right"></i>
+        </button>
+      </li>
+    </ul>
+  </nav>
+</div>
