+++ src/app/components/pagination/pagination.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PaginationComponent } from './pagination.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [PaginationComponent],
+  exports: [PaginationComponent]
+})
+export class PaginationModule { }
